+++ src/Autoloader.php
@@ -35,6 +35,12 @@ class Autoloader {
 			return self::$is_loaded;
 		}
 
+		
+		// If the main class has already been loaded, then they must be using a different autoloader.
+		if ( class_exists( 'WPGraphQL\RankMath\Main' ) ) {
+			return true;
+		}
+
 		$autoloader      = dirname( __DIR__ ) . '/vendor/autoload.php';
 		self::$is_loaded = self::require_autoloader( $autoloader );